Address CodeRabbit review feedback on graph-query workbench
- Relocate GraphQuery* and GraphSchema type definitions from
  src/types/graph-query.ts into the central src/types/index.ts to
  match the repo's shared type contract convention (CLAUDE.md).
- Escape Cypher identifiers when interpolating schema labels and
  relationship types into MATCH snippets, so names containing spaces,
  dashes, or backticks produce valid queries.
- Drive SchemaChip light/dark palette selection through
  useTheme().isDarkMode instead of reading document.documentElement
  at render time, so chips re-colour when the theme toggles.
